+++ src/main/java/com/raoulvdberge/refinedstorage/item/ItemWirelessGrid.java
@@ -1,5 +1,7 @@
 package com.raoulvdberge.refinedstorage.item;
 
+import baubles.api.BaubleType;
+import baubles.api.IBauble;
 import com.raoulvdberge.refinedstorage.RS;
 import com.raoulvdberge.refinedstorage.api.network.grid.IGrid;
 import com.raoulvdberge.refinedstorage.api.network.item.INetworkItem;
@@ -17,7 +19,7 @@ import net.minecraftforge.fml.relauncher.SideOnly;
 
 import javax.annotation.Nonnull;
 
-public class ItemWirelessGrid extends ItemNetworkItem {
+public class ItemWirelessGrid extends ItemNetworkItem implements IBauble {
     public ItemWirelessGrid(IItemInfo info) {
         super(info, RS.INSTANCE.config.wirelessGridCapacity);
     }
@@ -64,5 +66,10 @@ public class ItemWirelessGrid extends ItemNetworkItem {
 
     public static int getSize(ItemStack stack) {
         return (stack.hasTagCompound() && stack.getTagCompound().hasKey(NetworkNodeGrid.NBT_SIZE)) ? stack.getTagCompound().getInteger(NetworkNodeGrid.NBT_SIZE) : IGrid.SIZE_STRETCH;
+    }
+
+    @Override
+    public BaubleType getBaubleType(ItemStack itemstack) {
+        return BaubleType.AMULET;
     }
 }